Move COV_POOLS_FORMS_MAX_DECIMAL_PLACES to constants.js
